Extract balance summary tables
